requirements.txt
pyfiglet
pycrypto